backend/requirements.txt
# FastAPI and ASGI server
# Cache bust: 2025-12-19-v2 (force rebuild for shared folder fix)
fastapi==0.115.12
uvicorn[standard]==0.24.0
gunicorn==23.0.0  # Fixed: CVE-2024-1135, CVE-2024-6827 HTTP request smuggling
# Note: starlette version is managed by fastapi dependency
jinja2==3.1.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis and caching
redis==5.0.1
aioredis==2.0.1

# Authentication and security
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
cryptography==44.0.1
pyotp==2.9.0  # TOTP/HOTP for PACER MFA support

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.25.2
aiohttp>=3.12.14  # Fixed: GHSA-9548-qrrj-x5pj HTTP request smuggling
urllib3>=2.6.0  # Fixed: GHSA-gm62-xv2j-4w53, GHSA-2xpw-w6gg-jr37

# File processing
pypdf>=6.4.0  # Fixed: CVE LZWDecode RAM exhaustion vulnerability
# PyPDF2 removed - deprecated and has CVE infinite loop vulnerability, use pypdf instead
pdfplumber==0.10.3  # PDF text extraction for bankruptcy document processing
python-docx==1.1.0
python-magic==0.4.27
Pillow==11.1.0

# OCR and document processing
pytesseract==0.3.10
pdf2image==1.16.3

# AI and ML integrations
openai==1.58.1
anthropic>=0.40.0

# Natural language processing
spacy==3.7.2
nltk==3.9.1
eyecite==2.6.10

# Background tasks
celery==5.3.4
flower==2.0.1

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]>=2.20.0  # Fixed: CVE environment variable exposure to subprocesses
prometheus-client==0.19.0
psutil==5.9.6

# File storage
minio==7.2.0
boto3==1.34.0

# Utilities
python-dotenv==1.0.0
typer==0.9.0
rich==13.7.0
httptools==0.6.1
ujson==5.8.0

# Date and time
python-dateutil==2.8.2
pytz==2023.3

# Data processing
pandas==2.1.4
numpy>=1.26.0  # Python 3.12 requires numpy 1.26+ (pre-built wheels)

# XML/JSON processing
lxml==4.9.3
xmltodict==0.13.0

# Legal-specific libraries
# bluebook and legal-citation-parser removed - not available on PyPI

# Payment processing
stripe==7.8.0

# Email
aiosmtplib==3.0.1
sendgrid==6.11.0  # Optional: SendGrid email provider for marketing

# Validation and schema
jsonschema==4.20.0
marshmallow==3.20.1

# Security fixes for transitive dependencies
werkzeug>=3.1.4  # Fixed: GHSA-hgf8-39gv-g3f2

# Development and testing dependencies for optional dev builds
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==24.10.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1